[1/2] Decode CPDLC abortUser and abortProvider messages properly
These types of messages were so far displayed as unparsed, because
we required protectedMessage payload to be present, while abortUser
and abortProvider do not contain any payload. In this case the whole
outer object (ProtectedAircraftPDUs or ProtectedGroundPDUs) is
returned as a dissection result and we have to pass the dissected
message type back to the caller, because it may vary.

This creates a side effect of cmContactResponses being erroneously
dissected as abortProvider. Subsequent commit will fix this problem.