correlate_and_sync(): start searching for first preamble notch from 2 * SPS
So far we've been searching for the first notch of the preamble
starting at the sample, where the squelch has opened. On strong
transmissions squelch actually opens before the transmitter stable
state is reached. This caused the transmitter ramp-up stage
to be mistakenly identified as the first notch, which caused the
sync to fail. We therefore delay the starting point of the search
to two symbols past the squelch opening sample, when the transmitter
shall have reached stable state.