SDRPlay: remove gain setting in favor of gain reduction
Supported LNA gain reduction values differ between RSP
hardware types. This translates to different maxima of
gain setting in dumpvdl2 which is confusing. We
therefore drop the gain setting in favor of gain reduction.
20dB of GR means maximum gain, regardless of hardware type.
Upper limit of the --gr knob still varies across RSP types,
but this is not very important, as high GR values are not
very useful in VDL2 decoding.