Fix calculation of StatsD processing_time metric
Starting the timer in DEC_PREAMBLE phase means the result is biased,
because the frame has not yet been received from the air. We start
the timer in DEC_DATA phase instead - at this time the requested
number of bits has been already picked up from the air, so the metric
will only show the time spent on decoding. This is still slightly
biased, because the preamble searching and VDL2 header processing
is not included in the metric now. What's more, the metric includes
time spent on producing the output, which is not what we want.