Added protocol node type for raw octet strings
If a protocol is unknown or a PDU is unparseable, we want to print
it as raw hex. To avoid bloating the code of the formatters, it's
now possible to add such a PDU to the tree as a proto_DEF_unknown
node, which gets serialized into a hex string.