Graceful shutdown of thread on exit
Before termination we send a "shutdown" message to the decoder thread,
which is then proopagated to all output queues. Output drivers react to
this message by performing an orderly shutdown after the queue is
drained. The wait time is limited to 10 second, so that hung outputs do
not prevent the program from shutting down on signal.