Set the output instance state to enabled by default
Before this change, the output did not receive any messages before it
completed initialization. This may take a while (eg. when DNS resolution
is involved) and the output will receive nothing during that time. This
becomes apparent when decoding from a raw file - the program often
produces many messages rapidly and finishes quickly. Outputs that need
more time to initialize receive nothing in such cases. It is therefore
better to set the output state to enabled by default, hoping for the
initialization to complete successfully. Messages will then be queued up
instead of being dropped.