Reworked buggy COTP TPDU header length checking
Not only remaining buffer length must be sanity-checked, but the length
indicator as well.  Otherwise it may happen that we have enough space in
the buffer to parse variable header part, but the LI value is insanely
small.  If it's smaller than variable_part_offset, we end up passing a
negative (ie. insanely large postitive) length value to tlv_parse, which
in turn causes buffer overrun.