Add station id to statsd namespace
If the --station-id option is used, the station id gets appended to the "dumpvdl" namespace. For example, of you set the station id to KLAX, the new namespace would be dumpvdl2.KLAX. Adding this in case someone sets up multiple listeners and would like to monitor them seperately in Grafana. The changes are limited to statsd.c. I added an import of config.h and modified the statsd_initialize function.