reassembly: compute total PDU length automatically
Don't rely on PDU total length supplied in the fragment header. In CLNP
the total length includes the header of a full (uncompressed) PDU, which
we don't have, so we can't use this field to determine the total length
of a compressed PDU. This length is now calculated from the offset and
length of the final PDU fragment.