Separate offset-base reassembly engine from libacars
Currently there are two reassembly engine implementations:
- based on fragment offsets (in dumpvdl2, used only for CLNP)
- based on sequence numbers (in libacars, used for all other protocols)
Reassembly tables of these two engines use incompatible data structures,
so they need separate constructors/destructors and therefore can't be
stored in a common reassembly context. The interim solution is to
create two separate reassembly contexts - one for storing offset-based
reassembly tables and the other one - for sequence-based tables. We
store pointers to both context in a struct to avoid passing two
context pointers through the whole decoding stack of functions. This
is probably the least messy way to implement this for now, but it
needs an improvement.