fmtr-text: do not round up time
The current code rounds (up) milliseconds when dividing the
system-provided microsecond time. Furthermore, if the result of the
rounding exceeds 1000ms, it increments the system-provided seconds.

This is wrong as it effectively "pushes time forward", reporting a
future time that is not passed at the time of the timestamp collection.

Time should never be rounded, time scale divisions should always be
truncated to keep past, present and future in the right order.

This patch fixes this by removing the incorrect time adjustment.
tv_usec is guaranteed to always be in the range [0, 999,999]. This also
uses integer-based computation instead of (expensive) FP64.