calc_centerfreq(): adjust sampling bandwidth margin
Ideally the sampling rate should result in a spectrum bandwidth that
entirely covers the target frequencies plus additional spread to account
for the signal's own bandwidth.

This commit aligns the computation with that performed in calc_bandwidth()
(since they share the same rationale).

This is especially important in case a large sampling rate is used: e.g.
if using 10MSps, the previous implementation enforced a 2MHz margin
which is excessive.